+++ src/components/cell-tooltip/CellTooltipText.js
@@ -4,20 +4,22 @@ export default function CellTooltipText(props) {
   const {
     x,
     y,
+    parentWidth,
+    parentHeight,
     factors,
   } = props;
 
   const ref = useRef(null);
-  const offsetPercentage = 10;
-
-  // Do collision detection based on the bounds of the parent (.card-body) element.
+  const isNarrow = (parentWidth < 300);
+  // Do collision detection based on the bounds of the tooltip ancestor element.
   useEffect(() => {
     const el = ref.current;
-    const rect = el.parentNode.getBoundingClientRect();
-    const { width, height } = rect;
-    const translateX = (x > width / 2) ? -(100 + offsetPercentage) : offsetPercentage;
-    const translateY = (y > height / 2) ? -(100 + offsetPercentage) : offsetPercentage;
-    el.style.transform = `translateX(${translateX}%) translateY(${translateY}%)`;
+    const offsetPercentage = isNarrow ? -5 : 10;
+    const translateX = (x > parentWidth / 2) ? -(100 + offsetPercentage) : offsetPercentage;
+    const translateY = (y > parentHeight / 2) ? -(100 + offsetPercentage) : offsetPercentage;
+    const scale = isNarrow ? 0.75 : 1.0;
+    el.style.transform = `translateX(${translateX}%) translateY(${translateY}%) scale(${scale})`;
+    el.style.whiteSpace = (isNarrow ? 'normal' : 'nowrap');
   });
 
   return (