+++ src/components/classNames.js
@@ -1,4 +1,5 @@
-const CARD = 'card card-body my-2';
+export const TOOLTIP_ANCESTOR = 'tooltip-ancestor';
+const CARD = `card card-body my-2 ${TOOLTIP_ANCESTOR}`;
 export const BLACK_CARD = `${CARD} bg-black`;
 export const LIGHT_CARD = `${CARD} bg-light`;
 export const TITLE_CARD = 'title';